requirements.txt
google-transit-realtime-bindings==0.2.9
